requirements.txt
crewai[google-genai,litellm,tools]==1.14.1
crewai-tools
requests
beautifulsoup4
python-dotenv
streamlit
pydantic
google-auth
google-auth-oauthlib
google-api-python-client
google-genai
python-docx
